Started separating UI code
