+++ ProfilerView.cpp
