+++ TimelineTrack.h
@@ -16,6 +16,18 @@ namespace hop
 
 struct TimelineTrack
 {
+   TraceData _traces;
+   LockWaitData _lockWaits;
+   CoreEventData _coreEvents;
+   StrPtr_t _trackName{0};
+};
+
+size_t serializedSize( const TimelineTrack& ti );
+size_t serialize( const TimelineTrack& ti, char* );
+size_t deserialize( const char* data, TimelineTrack& ti );
+
+struct TimelineTrack
+{
    static float TRACE_HEIGHT;
    static float TRACE_VERTICAL_PADDING;
    static float PADDED_TRACE_SIZE;
@@ -32,9 +44,6 @@ struct TimelineTrack
    float heightWithThreadLabel() const noexcept;
    void setTrackHeight( float height );
    bool empty() const;
-   TraceData _traces;
-   LockWaitData _lockWaits;
-   CoreEventData _coreEvents;
 
    std::unordered_map< void*, std::vector< TimeStamp > > _lockWaitsPerMutex;
 
@@ -52,11 +61,6 @@ struct TimelineTrack
 
 private:
    float _trackHeight{9999.0f};
-   StrPtr_t _trackName{0};
-
-   friend size_t serializedSize( const TimelineTrack& ti );
-   friend size_t serialize( const TimelineTrack& ti, char* );
-   friend size_t deserialize( const char* data, TimelineTrack& ti );
 };
 
 struct LockOwnerInfo
