+++ common/Profiler.h
@@ -1,24 +1,12 @@
 #ifndef PROFILER_H_
 #define PROFILER_H_
 
-#include "Server.h"
-#include "Timeline.h"
-#include "TimelineTrack.h"
-#include "TraceSearch.h"
-#include "StringDb.h"
-
-#include <array>
-#include <chrono>
-#include <string>
-#include <vector>
-
 namespace hop
 {
-struct ProfilerStats;
-class Server;
+
 class Profiler
 {
-public:
+   friend class ProfilerView;
    enum SourceType
    {
       SRC_TYPE_NONE,
@@ -32,8 +20,6 @@ public:
    ProfilerStats stats() const;
    bool setSource( SourceType type, int processId, const char* str );
    SourceType sourceType() const;
-   void update( float deltaTimeMs, float globalTimeMs );
-   void draw( float drawPosX, float drawPosY, float windowWidth, float windowHeight );
    void fetchClientData();
    void addStringData( const std::vector< char >& stringData );
    void addTraces( const TraceData& traces, uint32_t threadIndex );
@@ -41,9 +27,6 @@ public:
    void addUnlockEvents(const std::vector<UnlockEvent>& unlockEvents, uint32_t threadIndex);
    void addCoreEvents( const std::vector<CoreEvent>& coreEvents, uint32_t threadIndex );
    void addThreadName( StrPtr_t name, uint32_t threadIndex );
-   void handleHotkey();
-   void handleMouse();
-   void setRecording( bool recording );
    void clear();
 
    bool saveToFile( const char* path );
@@ -54,7 +37,6 @@ private:
 
    std::string _name;
    int _pid;
-   Timeline _timeline;
    TimelineTracks _tracks;
    StringDb _strDb;
    bool _recording{ false };
@@ -64,14 +46,6 @@ private:
    Server::PendingData _serverPendingData;
 };
 
-struct ProfilerStats
-{
-   size_t strDbSize;
-   size_t traceCount;
-   size_t clientSharedMemSize;
-   int lodLevel;
-};
+}  // PROFILER_H_
 
-} // namespace hop
-
-#endif  // PROFILER_H_
+#endif // PROFILER_H_
